Add validation of shop ownership

+++ src/shop/shop.service.ts
@@ -1,15 +1,17 @@
-import { Injectable } from '@nestjs/common';
+import { ForbiddenException, Injectable } from '@nestjs/common';
 import { CreateShopDto } from './dto/create-shop.dto';
 import { UpdateShopDto } from './dto/update-shop.dto';
 import { Repository } from 'typeorm';
 import { Shop } from './entities/shop.entity';
 import { InjectRepository } from '@nestjs/typeorm';
+import { UserService } from '../user/user.service';
 
 @Injectable()
 export class ShopService {
   constructor(
     @InjectRepository(Shop)
     private readonly shopRepository: Repository<Shop>,
+    private readonly userService: UserService,
   ) {}
 
   async create(userId: number, createShopDto: CreateShopDto): Promise<Shop> {
@@ -31,10 +33,16 @@ export class ShopService {
     });
   }
 
-  async update(id: number, updateShopDto: UpdateShopDto): Promise<Shop> {
+  async update(
+    userId: number,
+    shopId: number,
+    updateShopDto: UpdateShopDto,
+  ): Promise<Shop> {
+    await this.validateShopOwner(userId, shopId);
+
     const shop = await this.shopRepository.findOneOrFail({
       where: {
-        id,
+        id: shopId,
       },
     });
 
@@ -46,12 +54,20 @@ export class ShopService {
     );
   }
 
-  async remove(id: number): Promise<Shop> {
+  async remove(userId: number, shopId: number): Promise<Shop> {
+    await this.validateShopOwner(userId, shopId);
+
     const shop = await this.shopRepository.findOneOrFail({
       where: {
-        id,
+        id: shopId,
       },
     });
     return await this.shopRepository.remove(shop);
+  }
+
+  async validateShopOwner(userId: number, shopId: number) {
+    const isMatch = await this.userService.validateShopOwner(userId, shopId);
+
+    if (!isMatch) throw new ForbiddenException('Invalid shop owner');
   }
 }
